core-java/core-java-datetime/src/main/java/io/github/kavahub/learnjava/use/UseToInstant.java
package io.github.kavahub.learnjava.use;

import java.time.LocalDateTime;
import java.time.ZoneId;
import java.util.Calendar;
import java.util.Date;

import lombok.experimental.UtilityClass;

@UtilityClass
public class UseToInstant {
    public LocalDateTime convertDateToLocalDate(Date date) {
        return LocalDateTime.ofInstant(date.toInstant(), ZoneId.systemDefault());
    }

    public LocalDateTime convertDateToLocalDate(Calendar calendar) {
        return LocalDateTime.ofInstant(calendar.toInstant(), ZoneId.systemDefault());
    }   
}
